refactor(dashboard-tsr): address CodeRabbit review findings (#1406-#1413)
Consolidated follow-up for the merged TanStack Start PRs:
- dedup: extract shared src/lib/api/clickhouse-config.ts (getClickHouseConfigsFromEnv
  + splitByComma) used by healthz/timezone/host-status/notifications (was copy-pasted)
- bug: cloudflare-workers-shim guards `typeof process` (was process.env at module top)
- naming: clerk-client isClerkEnabled -> isClerkClientEnabled (collided with env.ts's
  server-side isClerkEnabled; different semantics)

Build + prerender green on CF target. Skipped false-positives (raw-array API by
design; .filter(Boolean) matches Next behavior).
